Add pipeline messages for merge requests
Add support for webhooks for detached pipelines for merge requests.

If a pipeline belongs to a merge request (it has a non-null object for
the merge_request field in the webhook) then we look up pipeline and
merge request details from the GitLab API. This requires an API key from
the same GitLab instance that the webhook calls are coming from.